Improve the compatibility layer for Doctrine DBAL to avoid deprecations
