--- src/Tracing/Doctrine/DBAL/Compatibility/ExceptionConverterDriverInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility;
-
-/**
- * @internal
- */
-interface ExceptionConverterDriverInterface
-{
-}
